requirement.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
python-dotenv==1.0.0
coverage==7.2.3